read+delete Client Merge branch 'master' of https://github.com/salma-msalha/voyage-.git

+++ src/main/java/com/example/voyage/services/UserService.java
@@ -17,7 +17,7 @@ public interface UserService {
 
 	public Client updateClient(Client client);
 
-	public String deleteClient(int id);
+	public String deleteClient(Long id);
